refactor(proto): move the TRON bindings out from under dbfork
internal/dbfork/proto/ -> internal/tronproto/
scripts/gen-dbfork-protos.sh -> scripts/gen-tron-protos.sh

The package was placed under dbfork when dbfork was its only consumer.
Nothing in it is dbfork-specific — the Go package has always been called
`tronpb`, and the contents are upstream java-tron protos verbatim.

tools/txgen already imported it (node.go, for TransactionRaw), so the
mislabelling predates this branch. Adding the Wallet gRPC service stubs
made it acute: txgen's gRPC transport would be importing a gRPC client
from under the DB-mutation engine. Moving now costs one mechanical rename;
after the transport lands it would cost that plus a rewrite of the new
code's imports.

Pure rename — no generated bytes change. Verified by regenerating from the
new location and from `go generate ./internal/tronproto/...`: both leave
pb/ byte-identical, which is the same check the proto-drift CI job makes.

One real consequence, documented in the README rather than left to be
discovered: git subtree tracks by a `git-subtree-dir:` footer, and every
existing squash commit carries the pre-move path. A future
`git subtree pull --prefix=internal/tronproto/upstream` will not find the
previous split, so it cannot produce an incremental merge — the recorded
split sha is now in the README, along with a warning not to read an
empty-looking merge as "already up to date".

Also corrected the README's own framing, which still described the
directory as dbfork's and marked api/ as unused.
