+++ internal/dbfork/trc20_test.go
@@ -11,8 +11,8 @@ import (
 	"google.golang.org/protobuf/proto"
 
 	"github.com/tronprotocol/tron-deployment/internal/dbfork/db"
-	pb "github.com/tronprotocol/tron-deployment/internal/dbfork/proto/pb"
 	"github.com/tronprotocol/tron-deployment/internal/dbfork/stores"
+	pb "github.com/tronprotocol/tron-deployment/internal/tronproto/pb"
 )
 
 // TestKeccak256_EmptyVector pins that our keccak256 implementation is
