+++ internal/tronproto/upstream/core/contract/shield_contract.proto
