+++ internal/tronproto/upstream/core/contract/vote_asset_contract.proto
