+++ internal/tronproto/upstream/core/contract/witness_contract.proto
